Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 1680-concatenation-of-consecutive-binary-numbers/1680-concatenation-of-consecutive-binary-numbers.cpp
@@ -3,7 +3,7 @@ public:
 long long M=1e9+7;
     int concatenatedBinary(int n) {
         long long result=0;
-        long long digits=0;
+        int digits=0;
         for(int i=1;i<=n;i++){
             if((i&(i-1))==0)digits++;
             result=((result<<digits)%M+i)%M;
